requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
docutils==0.17.1
flake8==3.9.2
google-api-core==1.30.0
google-api-python-client==2.10.0
google-auth==1.32.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
jedi==0.17.2
Kivy==2.0.0
Kivy-Garden==0.1.4
kivymd==0.104.2
mccabe==0.6.1
oauthlib==3.1.1
packaging==20.9
parso==0.7.1
Pillow==8.2.0
pluggy==0.13.1
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.9.0
pyparsing==2.4.7
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
ujson==4.0.2
uritemplate==3.0.1
urllib3==1.26.6
yapf==0.31.0
